Users may now authorize with spotify
There are now 3 endpoints check-spotify, which is not implemented. My idea is that that function will check if the current user has stored spotify credentials and if so the button will be greyed out. Maybe there could be a remove button.

The other two endpoints are spotify-login and spotify-callback.

Spotify login initiates the process of reaching out to spotify and requesting access. If the user authenticates, spotify-callback is called. Spotify-callback also contains the logic for storing the refresh token for any given user. I decided to only store the refresh token because I felt storing the other info was kind of pointless, this may prove to be false, we shall see.

Row level security was added to the spotify credentials which means a token must be passed on supabase client creation that is valid and matches to a user. This might help prevent the table being dumped. I wanted to try it out it was pretty easy to implement in a small sql script which probably wont make it onto github because of datagrip and it's weirdnes. At least i just dont feel like fucking with it at this time.